+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Dev
 
 - Add `#[const(path::to::Trait::CONST)]` attribute to delegate associated constants via a getter.
+- Add `#[expr(<$ template>)]` attribute to modify delegated call using custom code.
 
 # 0.13.2 (14. 1. 2025)